SDN-3585: Add network tools imagestreams
The network tools image is used by certain tests in origin and the way
that is done today [1] is by spinning up a pod with the CVO image and
getting the network-tools image spec from inside the pod. This is not
desirable and also doesn't work today when these tests run on a cluster
with multi arch compute nodes.

This PR adds the network tools imagestreams so these tests can be
changed to use the imagestream rather than inferring it through creating
a pod today.

[1] https://github.com/openshift/origin/blob/master/test/extended/util/release.go#L32